Rename SoilGridsV2 -> SoilGrids2
