+++ src/DataWrangling/DataWrangling.jl
@@ -331,7 +331,7 @@ include("ORCA/ORCA.jl")
 include("WOA/WOA.jl")
 include("JRA55/JRA55.jl")
 include("OSPapa/OSPapa.jl")
-include("SoilGrids/SoilGridsV2.jl")
+include("SoilGrids/SoilGrids2.jl")
 include("IBCSO/IBCSO.jl")
 include("GEBCO/GEBCO.jl")
 include("IBCAO/IBCAO.jl")
